added entry-type, role, profession endpoints, employee endpoints renamed to user endpoints, adjusted the pattern of all endpoints

+++ src/main/java/com/pwc/ecasofond/repository/CompanyRepository.java
@@ -3,6 +3,6 @@ package com.pwc.ecasofond.repository;
 import com.pwc.ecasofond.model.Company;
 import org.springframework.data.repository.CrudRepository;
 
-public interface CompanyRepository extends CrudRepository<Company, Integer> {
+public interface CompanyRepository extends CrudRepository<Company, Long> {
 
 }
